+++ src/NSchema.Core/Model/SchemaAddress.cs
@@ -1,0 +1,14 @@
+namespace NSchema.Model;
+
+/// <summary>
+/// The address of a schema.
+/// </summary>
+/// <param name="Schema">The schema's name.</param>
+public sealed record SchemaAddress(SqlIdentifier Schema) : Address
+{
+    /// <inheritdoc />
+    public override string Value => Schema.Value;
+
+    /// <inheritdoc />
+    public override SqlIdentifier? SchemaName => null;
+}